add update and delete functionality

+++ src/controllers/books.controller.ts
@@ -1,8 +1,10 @@
 import { Request, Response, NextFunction } from "express";
 import httpStatus from "http-status";
+import { validationResult, Result } from "express-validator";
 
 import { BooksServices } from "../services";
 import { BookModel } from "../models/books.model.js";
+import { ApiError } from "../utils/apiError.js";
 import { IResData } from "./types";
 
 export class BooksControllers {
@@ -13,15 +15,28 @@ export class BooksControllers {
   }
 
   getBooks = async (
-    _req: Request,
+    req: Request,
     res: Response<IResData>,
     next: NextFunction
   ): Promise<any> => {
+    // const page = req.query.page as string;
+    // const limit = req.query.limit as string;
+    const { page, limit } = req.query as { page: string; limit: string };
     try {
-      const books = await this.booksService.finAllBooks();
-      console.log({ books });
-      res.status(httpStatus.OK).render("books", { books });
-      // res.status(httpStatus.OK).json({ status: httpStatus[200], data: books });
+      const result: Result = validationResult(req);
+      if (result.array().length > 0) {
+        throw new ApiError(httpStatus.BAD_REQUEST, result.array()[0]["msg"]);
+      }
+      const books = await this.booksService.findAllBooks({
+        page: +page,
+        limit: +limit,
+      });
+
+      res.status(httpStatus.OK).json({
+        status: httpStatus["200"],
+        message: "Get All Books!",
+        data: books,
+      });
     } catch (error) {
       next(error);
     }
@@ -33,12 +48,18 @@ export class BooksControllers {
     next: NextFunction
   ): Promise<any> => {
     const { id } = req.params;
-    console.log({ id });
     try {
+      const result: Result = validationResult(req);
+      if (result.array().length > 0) {
+        throw new ApiError(httpStatus.BAD_REQUEST, result.array()[0]["msg"]);
+      }
+
       const book = await this.booksService.finBookById(+id);
-      console.log({ book });
-      res.status(httpStatus.OK).render("book", { book });
-      // res.status(httpStatus.OK).json({ status: httpStatus["200"], book });
+      res.status(httpStatus.OK).json({
+        status: httpStatus["200"],
+        message: `Get Book with Id = ${id}`,
+        data: book,
+      });
     } catch (error) {
       next(error);
     }
@@ -50,8 +71,11 @@ export class BooksControllers {
     next: NextFunction
   ): Promise<void> => {
     const bookData = req.body;
-
     try {
+      const result: Result = validationResult(req);
+      if (result.array().length > 0) {
+        throw new ApiError(httpStatus.BAD_REQUEST, result.array()[0]["msg"]);
+      }
       const createdBook: BookModel = await this.booksService.createNewBook(
         bookData
       );
@@ -72,9 +96,14 @@ export class BooksControllers {
     next: NextFunction
   ): Promise<void> => {
     const bookData = req.body;
-    const id = req.params.id;
+    const { id } = req.params;
 
     try {
+      const result: Result = validationResult(req);
+      if (result.array().length > 0) {
+        throw new ApiError(httpStatus.BAD_REQUEST, result.array()[0]["msg"]);
+      }
+
       const updatedBook: BookModel = await this.booksService.updateBook(
         +id,
         bookData
@@ -95,10 +124,13 @@ export class BooksControllers {
     res: Response<IResData>,
     next: NextFunction
   ): Promise<void> => {
-    const bookData = req.body;
     const id = req.params.id;
 
     try {
+      const result: Result = validationResult(req);
+      if (result.array().length > 0) {
+        throw new ApiError(httpStatus.BAD_REQUEST, result.array()[0]["msg"]);
+      }
       await this.booksService.deleteBook(+id);
 
       res.status(httpStatus.OK).json({
